WICKET-6517 moved behavior out of markup package
This closes #256
